+++ python/mypy.ini
@@ -35,7 +35,7 @@ disallow_untyped_defs = False
 [mypy-pyspark.find_spark_home]
 disallow_untyped_defs = False
 
-[mypy-pyspark._globals]
+[mypy-pyspark.sql._globals]
 disallow_untyped_defs = False
 
 [mypy-pyspark.install]
